P8.6 stage 4b: the call-boundary law + the frame observers
The veto enforced: coordinates cross call boundaries AS Coordinates —
the splice unwrap dies. A coordinate-consuming fn (Style B) inlines
through the kernel rules, where f32/i32/extent/subscripts are legal
and the coordinate law still refuses arithmetic; coercing at the call
(Style A) keeps f a plain spliced scalar citizen — both sites legal,
the language never picks. extent(c) reads the frame's width (a host
int, build-time; the future extent-generic license's seat); f32/i32
widen to promote host ints explicitly (f32(extent(j))); f32 of a
float still refuses. Respells: caller-casts at the sites whose fns
capture data (the splice pins them); new committed spec tests for
Style B aspect-true normalization, the callee refusal, and the door
behaviors; 250 §3 records the boundary law and the deferred blanket
int-strictness.

+++ packages/tensorlib/tests/test_kernel.py
@@ -301,7 +301,7 @@ def test_all_scalar_subtrees_stay_value_dialect():
     @compute
     def apply_gain(f, img):
         (y,) = thread_idx("y")
-        img[y] = f(y)
+        img[y] = f(f32(y))
 
     img = T(np.zeros(3), ("y",))
     apply_gain(gain(2.0, 3.0), img)
@@ -329,7 +329,7 @@ def test_combinator_applied_twice_invalidates_its_sites():
     @compute
     def k(f, img):
         (y,) = thread_idx("y")
-        img[y] = f(y) + f(y)
+        img[y] = f(f32(y)) + f(f32(y))
 
     img = T(np.zeros(3), ("y",))
     sites = k.taps(scale(2.0), img)
@@ -499,7 +499,7 @@ def test_staged_transforms_compose_and_restage():
     def k(f, img):
         (y,) = thread_idx("y")
         g = slope_only(f)
-        v, (dy,) = g(y)
+        v, (dy,) = g(f32(y))
         img[y] = dy  # the slope field: constant a
 
     img = T(np.zeros(4), ("y",))
